requirements.txt
torch
torch-geometric
pyyaml
numpy
tqdm
